+++ build-logic/gobley-gradle-cargo/src/main/kotlin/dsl/DefaultCargoBuild.kt
@@ -46,5 +46,6 @@ abstract class DefaultCargoBuild<out RustTargetT : RustTarget, out CargoBuildVar
     init {
         @Suppress("LeakingThis")
         features.addAll(extension.features)
+        checkCommand.convention(extension.checkCommand)
     }
 }