+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Kotlin JVM and Kotlin Android Gradle plugins
   support ([#44](https://github.com/gobley/gobley/pull/44) & [#47](https://github.com/gobley/gobley/pull/47)).
 - Config option for using PascalCase for enums ([#54](https://github.com/gobley/gobley/pull/54)).
+- `CargoCheckTask` for cross-platform linting ([#55](https://github.com/gobley/gobley/pull/55)).
 
 ### Fixes